requirements.txt
streamlit==1.41.1
python-dotenv==1.1.1
openai==1.35.0

